Fix TextureRegions rendering upside down
Rendering TextureRegions using Graphics.drawTextureRegion rendered
them upside down as pointed in issue #35. I made a fix to Graphics
class that renders the texture region using negative heights so that
when invoking drawTextureRegion it renders as it should.

I still need to see if this creates a regression bug in Sprite's
behaviour due to it reversing already the TextureRegion.