Better support for XML namespace prefixes (#234)
Currently, if someone uses a prefix on the XML namespace (eg. <ui:container>) this would force them to redefine all the tag handling on the UiXmlLoader, which would not be desired. So instead we strip off the namespace prefix so we only deal with the real tag name so a user can name the prefix whatever they with and the UiXmlLoader should function as expected.